+++ cypress/selectors/viewDashboard.js
@@ -14,3 +14,7 @@ export const dashboardUnstarredSel = '[data-test="dashboard-unstarred"]'
 
 export const dragHandleSel = '[data-test="controlbar-drag-handle"]'
 export const dashboardsBarSel = '[data-test="dashboards-bar"]'
+
+export const outerScrollContainerSel = '[data-test="outer-scroll-container"]'
+
+export const editControlBarSel = '[data-test="edit-control-bar"]'