+++ src/api/systemSettings.js
@@ -1,5 +1,4 @@
 export const DEFAULT_SETTINGS = {
-    displayNameProperty: 'displayName',
     keyDashboardContextMenuItemOpenInRelevantApp: true,
     keyDashboardContextMenuItemShowInterpretationsAndDetails: true,
     keyDashboardContextMenuItemSwitchViewType: true,
@@ -16,11 +15,11 @@ const SYSTEM_SETTINGS = [
 ]
 
 const SYSTEM_SETTINGS_REMAPPINGS = {
-    keyDashboardContextMenuItemOpenInRelevantApp: 'openInRelevantApp',
+    keyDashboardContextMenuItemOpenInRelevantApp: 'allowVisOpenInApp',
     keyDashboardContextMenuItemShowInterpretationsAndDetails:
-        'showInterpretationsAndDetails',
-    keyDashboardContextMenuItemSwitchViewType: 'switchViewType',
-    keyDashboardContextMenuItemViewFullscreen: 'fullscreenAllowedInSettings',
+        'allowVisShowInterpretations',
+    keyDashboardContextMenuItemSwitchViewType: 'allowVisViewAs',
+    keyDashboardContextMenuItemViewFullscreen: 'allowVisFullscreen',
 }
 
 export const renameSystemSettings = settings => {
@@ -30,9 +29,7 @@ export const renameSystemSettings = settings => {
     }, {})
 }
 
-const query = {
+export const systemSettingsQuery = {
     resource: 'systemSettings',
     params: { key: SYSTEM_SETTINGS },
 }
-
-export default query