+++ src/components/ControlBar/EditBar.js
@@ -8,24 +8,29 @@ import { Button, ButtonStrip } from '@dhis2/ui'
 import { useDataEngine, useAlert } from '@dhis2/app-runtime'
 import { useD2 } from '@dhis2/app-runtime-adapter-d2'
 
+import FilterSettingsDialog from '../ItemFilter/FilterSettingsDialog'
 import ConfirmDeleteDialog from './ConfirmDeleteDialog'
+import { useUserSettings } from '../UserSettingsProvider'
 import {
     tSaveDashboard,
     acClearEditDashboard,
     acSetPrintPreviewView,
     acClearPrintPreviewView,
+    acSetFilterSettings,
 } from '../../actions/editDashboard'
 import { acClearPrintDashboard } from '../../actions/printDashboard'
 import {
     tDeleteDashboard,
     acSetDashboardDisplayName,
 } from '../../actions/dashboards'
+import { sGetDimensions } from '../../reducers/dimensions'
 import {
     sGetEditDashboardRoot,
     sGetIsNewDashboard,
     sGetIsPrintPreviewView,
 } from '../../reducers/editDashboard'
-import { apiFetchDashboard } from '../../api/dashboards'
+import { apiFetchDashboard } from '../../api/fetchDashboard'
+import { EDIT } from '../Dashboard/dashboardModes'
 
 import classes from './styles/EditBar.module.css'
 
@@ -34,12 +39,17 @@ const saveFailedMessage = i18n.t(
 )
 
 const EditBar = props => {
-    const { d2 } = useD2({})
+    const { d2 } = useD2()
     const dataEngine = useDataEngine()
     const [translationDlgIsOpen, setTranslationDlgIsOpen] = useState(false)
+    const [filterSettingsDlgIsOpen, setFilterSettingsDlgIsOpen] = useState(
+        false
+    )
     const [dashboard, setDashboard] = useState(undefined)
     const [confirmDeleteDlgIsOpen, setConfirmDeleteDlgIsOpen] = useState(false)
     const [redirectUrl, setRedirectUrl] = useState(undefined)
+
+    const { userSettings } = useUserSettings()
 
     const failureAlert = useAlert(saveFailedMessage, {
         critical: true,
@@ -47,9 +57,11 @@ const EditBar = props => {
 
     useEffect(() => {
         if (props.dashboardId && !dashboard) {
-            apiFetchDashboard(dataEngine, props.dashboardId).then(dboard =>
-                setDashboard(dboard)
-            )
+            apiFetchDashboard(
+                dataEngine,
+                props.dashboardId,
+                EDIT
+            ).then(dboard => setDashboard(dboard))
         }
     }, [props.dashboardId, dashboard])
 
@@ -93,14 +105,26 @@ const EditBar = props => {
         })
     }
 
-    const onTranslationsSaved = async translations => {
-        if (translations && translations.length) {
-            const dbLocale = await d2.currentUser.userSettings.get(
-                'keyDbLocale'
-            )
+    const onFilterSettingsConfirmed = (
+        filterSettingsRestrictability,
+        selectedFilters
+    ) => {
+        const allowedFilters = filterSettingsRestrictability
+            ? selectedFilters
+            : []
+        props.setFilterSettings({
+            allowedFilters,
+            restrictFilters: filterSettingsRestrictability,
+        })
+        toggleFilterSettingsDialog()
+    }
 
+    const onTranslationsSaved = translations => {
+        if (translations && translations.length) {
             const translation = translations.find(
-                t => t.locale === dbLocale && t.property === 'NAME'
+                t =>
+                    t.locale === userSettings.keyDbLocale &&
+                    t.property === 'NAME'
             )
 
             if (translation && translation.value) {
@@ -111,6 +135,10 @@ const EditBar = props => {
 
     const toggleTranslationDialog = () =>
         setTranslationDlgIsOpen(!translationDlgIsOpen)
+
+    const toggleFilterSettingsDialog = () => {
+        setFilterSettingsDlgIsOpen(!filterSettingsDlgIsOpen)
+    }
 
     const confirmDeleteDialog = () =>
         props.deleteAccess && props.dashboardId ? (
@@ -141,6 +169,18 @@ const EditBar = props => {
             />
         ) : null
 
+    const filterSettingsDialog = () =>
+        dashboard || props.newDashboard ? (
+            <FilterSettingsDialog
+                dimensions={props.dimensions}
+                restrictFilters={props.restrictFilters}
+                initiallySelectedItems={props.allowedFilters}
+                onClose={toggleFilterSettingsDialog}
+                onConfirm={onFilterSettingsConfirmed}
+                open={filterSettingsDlgIsOpen}
+            />
+        ) : null
+
     const renderActionButtons = () => (
         <ButtonStrip>
             <Button primary onClick={onSave} dataTest="save-dashboard-button">
@@ -151,7 +191,9 @@ const EditBar = props => {
                     ? i18n.t('Exit Print preview')
                     : i18n.t('Print preview')}
             </Button>
-
+            <Button onClick={toggleFilterSettingsDialog}>
+                {i18n.t('Filter settings')}
+            </Button>
             {props.dashboardId && (
                 <Button onClick={toggleTranslationDialog}>
                     {i18n.t('Translate')}
@@ -172,22 +214,21 @@ const EditBar = props => {
         return <Redirect to={redirectUrl} />
     }
 
-    const { updateAccess } = props
-
-    const discardBtnText = updateAccess
+    const discardBtnText = props.updateAccess
         ? i18n.t('Exit without saving')
         : i18n.t('Go to dashboards')
 
     return (
         <>
-            <div className={classes.editBar}>
+            <div className={classes.editBar} data-test="edit-control-bar">
                 <div className={classes.controls}>
-                    {updateAccess ? renderActionButtons() : null}
+                    {props.updateAccess ? renderActionButtons() : null}
                     <Button secondary onClick={onDiscard}>
                         {discardBtnText}
                     </Button>
                 </div>
             </div>
+            {filterSettingsDialog()}
             {translationDialog()}
             {confirmDeleteDialog()}
         </>
@@ -195,13 +236,18 @@ const EditBar = props => {
 }
 
 EditBar.propTypes = {
+    allowedFilters: PropTypes.array,
     clearPrintDashboard: PropTypes.func,
     clearPrintPreview: PropTypes.func,
     dashboardId: PropTypes.string,
     dashboardName: PropTypes.string,
     deleteAccess: PropTypes.bool,
+    dimensions: PropTypes.array,
     isPrintPreviewView: PropTypes.bool,
+    newDashboard: PropTypes.bool,
+    restrictFilters: PropTypes.bool,
     saveDashboard: PropTypes.func,
+    setFilterSettings: PropTypes.func,
     showPrintPreview: PropTypes.func,
     updateAccess: PropTypes.bool,
     onDelete: PropTypes.func,
@@ -211,21 +257,27 @@ EditBar.propTypes = {
 
 const mapStateToProps = state => {
     const dashboard = sGetEditDashboardRoot(state)
-
+    let newDashboard
     let deleteAccess
     let updateAccess
     if (sGetIsNewDashboard(state)) {
+        newDashboard = true
         deleteAccess = true
         updateAccess = true
     } else {
+        newDashboard = false
         updateAccess = dashboard.access ? dashboard.access.update : false
         deleteAccess = dashboard.access ? dashboard.access.delete : false
     }
 
     return {
+        allowedFilters: dashboard.allowedFilters,
         dashboardId: dashboard.id,
         dashboardName: dashboard.name,
         deleteAccess,
+        dimensions: sGetDimensions(state),
+        newDashboard,
+        restrictFilters: dashboard.restrictFilters,
         isPrintPreviewView: sGetIsPrintPreviewView(state),
         updateAccess,
     }
@@ -238,6 +290,7 @@ const mapDispatchToProps = dispatch => ({
     onDelete: id => dispatch(tDeleteDashboard(id)),
     onDiscardChanges: () => dispatch(acClearEditDashboard()),
     onTranslate: (id, value) => dispatch(acSetDashboardDisplayName(id, value)),
+    setFilterSettings: value => dispatch(acSetFilterSettings(value)),
     showPrintPreview: () => dispatch(acSetPrintPreviewView()),
 })