+++ src/components/Dashboard/ViewDashboard.js
@@ -11,8 +11,11 @@ import DashboardsBar from '../ControlBar/ViewControlBar/DashboardsBar'
 import { sGetIsEditing } from '../../reducers/editDashboard'
 import { sGetIsPrinting } from '../../reducers/printDashboard'
 import { sGetSelectedId } from '../../reducers/selected'
+import { sGetPassiveViewRegistered } from '../../reducers/passiveViewRegistered'
 import { acClearEditDashboard } from '../../actions/editDashboard'
 import { acClearPrintDashboard } from '../../actions/printDashboard'
+import { acSetPassiveViewRegistered } from '../../actions/passiveViewRegistered'
+import { apiPostDataStatistics } from '../../api/dataStatistics'
 
 import classes from './styles/ViewDashboard.module.css'
 
@@ -35,10 +38,24 @@ export const ViewDashboard = props => {
         })
     }, [props.selectedId])
 
+    useEffect(() => {
+        if (!props.passiveViewRegistered) {
+            apiPostDataStatistics(
+                'PASSIVE_DASHBOARD_VIEW',
+                props.selectedId
+            ).then(() => {
+                props.registerPassiveView()
+            })
+        }
+    }, [props.passiveViewRegistered])
+
     const onExpandedChanged = expanded => setControlbarExpanded(expanded)
 
     return (
-        <div className={cx(classes.container, 'dashboard-scroll-container')}>
+        <div
+            className={cx(classes.container, 'dashboard-scroll-container')}
+            data-test="outer-scroll-container"
+        >
             <DashboardsBar
                 expanded={controlbarExpanded}
                 onExpandedChanged={onExpandedChanged}
@@ -64,6 +81,8 @@ ViewDashboard.propTypes = {
     clearPrintDashboard: PropTypes.func,
     dashboardIsEditing: PropTypes.bool,
     dashboardIsPrinting: PropTypes.bool,
+    passiveViewRegistered: PropTypes.bool,
+    registerPassiveView: PropTypes.func,
     selectedId: PropTypes.string,
 }
 
@@ -71,6 +90,7 @@ const mapStateToProps = state => {
     return {
         dashboardIsEditing: sGetIsEditing(state),
         dashboardIsPrinting: sGetIsPrinting(state),
+        passiveViewRegistered: sGetPassiveViewRegistered(state),
         selectedId: sGetSelectedId(state),
     }
 }
@@ -78,4 +98,5 @@ const mapStateToProps = state => {
 export default connect(mapStateToProps, {
     clearEditDashboard: acClearEditDashboard,
     clearPrintDashboard: acClearPrintDashboard,
+    registerPassiveView: acSetPassiveViewRegistered,
 })(ViewDashboard)