+++ src/components/Item/VisualizationItem/Visualization/__tests__/Visualization.spec.js
@@ -2,6 +2,7 @@ import React from 'react'
 import { render } from '@testing-library/react'
 import { Provider } from 'react-redux'
 import configureMockStore from 'redux-mock-store'
+import UserSettingsProvider from '../../../../UserSettingsProvider'
 
 import Visualization from '../Visualization'
 
@@ -39,20 +40,27 @@ const DEFAULT_STORE_WITH_ONE_ITEM = {
 
 global.eventChartPlugin = {}
 global.eventReportPlugin = {}
+global.fetch = jest.fn(() =>
+    Promise.resolve({
+        userSettings: { keyAnalysisDisplayPropert: 'name' },
+    })
+)
 
 test('renders a MapPlugin when activeType is MAP', () => {
     const { container } = render(
         <Provider store={mockStore(DEFAULT_STORE_WITH_ONE_ITEM)}>
-            <Visualization
-                item={{
-                    id: 'rainbow',
-                    type: 'MAP',
-                    map: { id: 'rainbowVis' },
-                }}
-                activeType="MAP"
-                itemFilters={{}}
-                availableHeight={500}
-            />
+            <UserSettingsProvider>
+                <Visualization
+                    item={{
+                        id: 'rainbow',
+                        type: 'MAP',
+                        map: { id: 'rainbowVis' },
+                    }}
+                    activeType="MAP"
+                    itemFilters={{}}
+                    availableHeight={500}
+                />
+            </UserSettingsProvider>
         </Provider>
     )
     expect(container).toMatchSnapshot()
@@ -61,16 +69,18 @@ test('renders a MapPlugin when activeType is MAP', () => {
 test('renders a VisualizationPlugin for CHART', () => {
     const { container } = render(
         <Provider store={mockStore(DEFAULT_STORE_WITH_ONE_ITEM)}>
-            <Visualization
-                item={{
-                    id: 'rainbow',
-                    type: 'CHART',
-                    chart: { id: 'rainbowVis' },
-                }}
-                activeType="CHART"
-                itemFilters={{}}
-                availableHeight={500}
-            />
+            <UserSettingsProvider>
+                <Visualization
+                    item={{
+                        id: 'rainbow',
+                        type: 'CHART',
+                        chart: { id: 'rainbowVis' },
+                    }}
+                    activeType="CHART"
+                    itemFilters={{}}
+                    availableHeight={500}
+                />
+            </UserSettingsProvider>
         </Provider>
     )
     expect(container).toMatchSnapshot()
@@ -79,16 +89,18 @@ test('renders a VisualizationPlugin for CHART', () => {
 test('renders a VisualizationPlugin for REPORT_TABLE', () => {
     const { container } = render(
         <Provider store={mockStore(DEFAULT_STORE_WITH_ONE_ITEM)}>
-            <Visualization
-                item={{
-                    id: 'rainbow',
-                    type: 'REPORT_TABLE',
-                    reportTable: { id: 'rainbowVis' },
-                }}
-                activeType="REPORT_TABLE"
-                itemFilters={{}}
-                availableHeight={500}
-            />
+            <UserSettingsProvider>
+                <Visualization
+                    item={{
+                        id: 'rainbow',
+                        type: 'REPORT_TABLE',
+                        reportTable: { id: 'rainbowVis' },
+                    }}
+                    activeType="REPORT_TABLE"
+                    itemFilters={{}}
+                    availableHeight={500}
+                />
+            </UserSettingsProvider>
         </Provider>
     )
     expect(container).toMatchSnapshot()
@@ -97,16 +109,18 @@ test('renders a VisualizationPlugin for REPORT_TABLE', () => {
 test('renders active type MAP rather than original type REPORT_TABLE', () => {
     const { container } = render(
         <Provider store={mockStore(DEFAULT_STORE_WITH_ONE_ITEM)}>
-            <Visualization
-                item={{
-                    id: 'rainbow',
-                    type: 'REPORT_TABLE',
-                    reportTable: { id: 'rainbowVis' },
-                }}
-                activeType="MAP"
-                itemFilters={{}}
-                availableHeight={500}
-            />
+            <UserSettingsProvider>
+                <Visualization
+                    item={{
+                        id: 'rainbow',
+                        type: 'REPORT_TABLE',
+                        reportTable: { id: 'rainbowVis' },
+                    }}
+                    activeType="MAP"
+                    itemFilters={{}}
+                    availableHeight={500}
+                />
+            </UserSettingsProvider>
         </Provider>
     )
     expect(container).toMatchSnapshot()
@@ -115,16 +129,18 @@ test('renders active type MAP rather than original type REPORT_TABLE', () => {
 test('renders active type REPORT_TABLE rather than original type MAP', () => {
     const { container } = render(
         <Provider store={mockStore(DEFAULT_STORE_WITH_ONE_ITEM)}>
-            <Visualization
-                item={{
-                    id: 'rainbow',
-                    type: 'MAP',
-                    map: { id: 'rainbowVis' },
-                }}
-                activeType="REPORT_TABLE"
-                itemFilters={{}}
-                availableHeight={500}
-            />
+            <UserSettingsProvider>
+                <Visualization
+                    item={{
+                        id: 'rainbow',
+                        type: 'MAP',
+                        map: { id: 'rainbowVis' },
+                    }}
+                    activeType="REPORT_TABLE"
+                    itemFilters={{}}
+                    availableHeight={500}
+                />
+            </UserSettingsProvider>
         </Provider>
     )
     expect(container).toMatchSnapshot()
@@ -133,16 +149,18 @@ test('renders active type REPORT_TABLE rather than original type MAP', () => {
 test('renders a DefaultPlugin when activeType is EVENT_CHART', () => {
     const { container } = render(
         <Provider store={mockStore(DEFAULT_STORE_WITH_ONE_ITEM)}>
-            <Visualization
-                item={{
-                    id: 'rainbow',
-                    type: 'EVENT_CHART',
-                    chart: { id: 'rainbowVis' },
-                }}
-                activeType="EVENT_CHART"
-                itemFilters={{}}
-                availableHeight={500}
-            />
+            <UserSettingsProvider>
+                <Visualization
+                    item={{
+                        id: 'rainbow',
+                        type: 'EVENT_CHART',
+                        chart: { id: 'rainbowVis' },
+                    }}
+                    activeType="EVENT_CHART"
+                    itemFilters={{}}
+                    availableHeight={500}
+                />
+            </UserSettingsProvider>
         </Provider>
     )
     expect(container).toMatchSnapshot()
@@ -151,16 +169,18 @@ test('renders a DefaultPlugin when activeType is EVENT_CHART', () => {
 test('renders a DefaultPlugin when activeType is EVENT_REPORT', () => {
     const { container } = render(
         <Provider store={mockStore(DEFAULT_STORE_WITH_ONE_ITEM)}>
-            <Visualization
-                item={{
-                    id: 'rainbow',
-                    type: 'EVENT_REPORT',
-                    chart: { id: 'rainbowVis' },
-                }}
-                activeType="EVENT_REPORT"
-                itemFilters={{}}
-                availableHeight={500}
-            />
+            <UserSettingsProvider>
+                <Visualization
+                    item={{
+                        id: 'rainbow',
+                        type: 'EVENT_REPORT',
+                        chart: { id: 'rainbowVis' },
+                    }}
+                    activeType="EVENT_REPORT"
+                    itemFilters={{}}
+                    availableHeight={500}
+                />
+            </UserSettingsProvider>
         </Provider>
     )
     expect(container).toMatchSnapshot()
@@ -172,12 +192,14 @@ test('renders NoVisMessage when no visualization', () => {
     }
     const { container } = render(
         <Provider store={mockStore(store)}>
-            <Visualization
-                item={{ id: 'rainbow' }}
-                activeType="CHART"
-                itemFilters={{}}
-                availableHeight={500}
-            />
+            <UserSettingsProvider>
+                <Visualization
+                    item={{ id: 'rainbow' }}
+                    activeType="CHART"
+                    itemFilters={{}}
+                    availableHeight={500}
+                />
+            </UserSettingsProvider>
         </Provider>
     )
     expect(container).toMatchSnapshot()