+++ src/components/__tests__/App.spec.js
@@ -4,13 +4,13 @@ import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
 import { render } from '@testing-library/react'
 import { useD2 } from '@dhis2/app-runtime-adapter-d2'
-import { useSystemSettings } from '../SystemSettingsProvider'
+import { useUserSettings } from '../UserSettingsProvider'
 import { apiFetchDimensions } from '@dhis2/analytics'
-import { apiFetchDashboards } from '../../api/dashboards'
+import { apiFetchDashboards } from '../../api/fetchAllDashboards'
 import App from '../App'
 
-jest.mock('../../api/dashboards')
-jest.mock('../SystemSettingsProvider')
+jest.mock('../../api/fetchAllDashboards')
+jest.mock('../UserSettingsProvider')
 jest.mock('@dhis2/analytics')
 jest.mock('@dhis2/app-runtime-adapter-d2')
 jest.mock(
@@ -30,17 +30,14 @@ test('renders the app', () => {
             currentUser: 'rainbowDash',
         },
     })
-    useSystemSettings.mockReturnValue({
-        settings: { displayNameProperty: 'displayName' },
+    useUserSettings.mockReturnValue({
+        userSettings: { keyAnalysisDisplayProperty: 'displayName' },
     })
     apiFetchDashboards.mockReturnValue([
         {
             id: 'rainbowdash',
-            name: 'Rainbow Dash',
-            dashboardItems: [],
-            user: {},
-            created: 'today',
-            lastUpdated: 'today',
+            displayName: 'Rainbow Dash',
+            starred: true,
         },
     ])
     apiFetchDimensions.mockReturnValue([{ dimensionType: 'mock' }])